requirements.txt
# Telegram Bot Dependencies
telethon==1.32.1
python-telegram-bot==20.7
pyrogram==2.0.106
tgcrypto==1.2.5

# Google Sheets Integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gspread==5.12.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2

# Environment & Configuration
python-dotenv==1.0.0
configparser==6.0.0

# Utilities
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
Pillow==10.1.0

# Async Support
aiohttp==3.9.1
asyncio-throttle==1.0.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
